pyproject.toml
[project]
name = "uk-resell-adk"
version = "0.1.0"
description = "Google ADK multi-agent system for sourcing, evaluating, and reporting UK resale opportunities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "google-adk>=0.3.0",
  "httpx>=0.27.0",
  "langfuse>=3.0.0",
  "langsmith>=0.1.0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0",
  "opentelemetry-sdk>=1.27.0",
  "psycopg[binary]>=3.2.0",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
